ai/viettune-whisper-service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
openai-whisper @ git+https://github.com/openai/whisper.git
torch>=2.0.0
python-dotenv==1.0.1
pydantic-settings==2.7.0
setuptools
wheel